refactor: extract types.ts
